src/main/java/com/n26/statistics/StatisticsController.java
package com.n26.statistics;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController()
@RequestMapping("/statistics")
public class StatisticsController {

    @Autowired
    private StatisticsService statisticsService;

    @GetMapping
    public StatisticsResponse getStatistics(){
        Statistics statistics = this.statisticsService.getStatistics();

        return StatisticsResponse.fromStatistics(statistics);
    }

}
